Split prototype into separate components
